add conf connection to themer module
